refactor(hbgw): move Gateway validation functions to refs package
Move GetSupportedGatewayForParentRef from route to refs package
for better code organization and reusability.
