+++ controller/hybridgateway/route/status_test.go
@@ -2,7 +2,6 @@ package route
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"reflect"
 	"strings"
@@ -20,10 +19,7 @@ import (
 	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
 	gatewayv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 
-	hybridgatewayerrors "github.com/kong/kong-operator/controller/hybridgateway/errors"
-
 	gwtypes "github.com/kong/kong-operator/internal/types"
-	"github.com/kong/kong-operator/pkg/vars"
 )
 
 // Test helpers for BuildProgrammedCondition
@@ -549,178 +545,6 @@ func (e *errorClient) Get(ctx context.Context, key client.ObjectKey, obj client.
 		return fmt.Errorf("generic gatewayclass error")
 	}
 	return e.Client.Get(ctx, key, obj, opts...)
-}
-
-func Test_GetSupportedGatewayForParentRef(t *testing.T) {
-	ctx := context.Background()
-	logger := logr.Discard()
-
-	controllerName := vars.DefaultControllerName
-	vars.SetControllerName(controllerName)
-
-	s := runtime.NewScheme()
-	_ = gatewayv1.Install(s)
-
-	gateway := &gwtypes.Gateway{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: "default",
-			Name:      "my-gateway",
-		},
-		Spec: gwtypes.GatewaySpec{
-			GatewayClassName: "my-class",
-		},
-	}
-	gateway.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "gateway.networking.k8s.io",
-		Version: "v1",
-		Kind:    "Gateway",
-	})
-	gatewayClass := &gwtypes.GatewayClass{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "my-class",
-		},
-		Spec: gwtypes.GatewayClassSpec{
-			ControllerName: gwtypes.GatewayController(controllerName),
-		},
-	}
-	gatewayClass.SetGroupVersionKind(schema.GroupVersionKind{
-		Group:   "gateway.networking.k8s.io",
-		Version: "v1",
-		Kind:    "GatewayClass",
-	})
-
-	tests := []struct {
-		name          string
-		pRef          gwtypes.ParentReference
-		routeNS       string
-		objs          []client.Object
-		controllerVal string
-		wantErr       error
-		wantNil       bool
-	}{
-		{
-			name:    "unsupported kind",
-			pRef:    gwtypes.ParentReference{Kind: kindPtr("OtherKind"), Name: "my-gateway"},
-			routeNS: "default",
-			objs:    []client.Object{gateway, gatewayClass},
-			wantNil: true,
-		},
-		{
-			name:    "unsupported group",
-			pRef:    gwtypes.ParentReference{Kind: kindPtr("Gateway"), Group: groupPtr("other.group"), Name: "my-gateway"},
-			routeNS: "default",
-			objs:    []client.Object{gateway, gatewayClass},
-			wantNil: true,
-		},
-		{
-			name:    "gateway not found",
-			pRef:    gwtypes.ParentReference{Kind: kindPtr("Gateway"), Group: groupPtr("gateway.networking.k8s.io"), Name: "notfound"},
-			routeNS: "default",
-			objs:    []client.Object{},
-			wantErr: fmt.Errorf("no supported gateway found"),
-		},
-		{
-			name:    "gateway class not found",
-			pRef:    gwtypes.ParentReference{Kind: kindPtr("Gateway"), Group: groupPtr("gateway.networking.k8s.io"), Name: "my-gateway"},
-			routeNS: "default",
-			objs:    []client.Object{gateway},
-			wantErr: fmt.Errorf("no gatewayClass found for gateway"),
-		},
-		{
-			name:    "gateway class wrong controller",
-			pRef:    gwtypes.ParentReference{Kind: kindPtr("Gateway"), Group: groupPtr("gateway.networking.k8s.io"), Name: "my-gateway"},
-			routeNS: "default",
-			objs: []client.Object{gateway, &gwtypes.GatewayClass{
-				ObjectMeta: metav1.ObjectMeta{Name: "my-class"},
-				Spec:       gwtypes.GatewayClassSpec{ControllerName: "wrong-controller"},
-			}},
-			wantErr: fmt.Errorf("gatewayClass is not controlled by this controller"),
-		},
-		{
-			name:    "gateway class empty controller",
-			pRef:    gwtypes.ParentReference{Kind: kindPtr("Gateway"), Group: groupPtr("gateway.networking.k8s.io"), Name: "my-gateway"},
-			routeNS: "default",
-			objs: []client.Object{gateway, &gwtypes.GatewayClass{
-				ObjectMeta: metav1.ObjectMeta{Name: "my-class"},
-				Spec:       gwtypes.GatewayClassSpec{ControllerName: ""},
-			}},
-			wantErr: fmt.Errorf("gatewayClass is not controlled by this controller"),
-		},
-		{
-			name:    "supported parent ref",
-			pRef:    gwtypes.ParentReference{Kind: kindPtr("Gateway"), Group: groupPtr("gateway.networking.k8s.io"), Name: "my-gateway"},
-			routeNS: "default",
-			objs:    []client.Object{gateway, gatewayClass},
-			wantNil: false,
-		},
-		{
-			name:    "gateway get generic error",
-			pRef:    gwtypes.ParentReference{Kind: kindPtr("Gateway"), Group: groupPtr("gateway.networking.k8s.io"), Name: "my-gateway"},
-			routeNS: "default",
-			objs:    []client.Object{gateway, gatewayClass},
-			wantErr: fmt.Errorf("failed to get gateway for ParentReference"),
-		},
-		{
-			name:    "gatewayclass get generic error",
-			pRef:    gwtypes.ParentReference{Kind: kindPtr("Gateway"), Group: groupPtr("gateway.networking.k8s.io"), Name: "my-gateway"},
-			routeNS: "default",
-			objs:    []client.Object{gateway, gatewayClass},
-			wantErr: fmt.Errorf("failed to get gatewayClass for parentReference"),
-		},
-		{
-			name:    "parentRef with custom namespace",
-			pRef:    gwtypes.ParentReference{Kind: kindPtr("Gateway"), Group: groupPtr("gateway.networking.k8s.io"), Name: "my-gateway", Namespace: nsPtr("custom-ns")},
-			routeNS: "default",
-			objs: []client.Object{
-				&gwtypes.Gateway{
-					ObjectMeta: metav1.ObjectMeta{
-						Namespace: "custom-ns",
-						Name:      "my-gateway",
-					},
-					Spec: gwtypes.GatewaySpec{
-						GatewayClassName: "my-class",
-					},
-				},
-				&gwtypes.GatewayClass{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "my-class",
-					},
-					Spec: gwtypes.GatewayClassSpec{
-						ControllerName: gwtypes.GatewayController(vars.DefaultControllerName),
-					},
-				},
-			},
-			wantNil: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			baseClient := fake.NewClientBuilder().WithScheme(s).WithObjects(tt.objs...).Build()
-			var cl client.Client = baseClient
-			if tt.name == "gateway get generic error" {
-				cl = &errorClient{Client: baseClient, failGateway: true}
-			}
-			if tt.name == "gatewayclass get generic error" {
-				cl = &errorClient{Client: baseClient, failGatewayClass: true}
-			}
-			gw, err := GetSupportedGatewayForParentRef(ctx, logger, cl, tt.pRef, tt.routeNS)
-			if tt.wantErr != nil {
-				require.Error(t, err)
-				if errors.Is(err, hybridgatewayerrors.ErrNoGatewayFound) || errors.Is(err, hybridgatewayerrors.ErrNoGatewayClassFound) || errors.Is(err, hybridgatewayerrors.ErrNoGatewayController) {
-					// Specific error type matches
-					return
-				}
-				require.Contains(t, err.Error(), tt.wantErr.Error())
-				return
-			}
-			if tt.wantNil {
-				require.Nil(t, gw)
-			} else {
-				require.NotNil(t, gw)
-			}
-		})
-	}
 }
 
 func Test_BuildAcceptedCondition(t *testing.T) {
